requirements.txt
# SOC Triage System Dependencies

# HTTP client for API calls
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Vector database for RAG
chromadb>=0.4.0

# LangChain for RAG pipeline (optional, for direct integration)
# langchain>=0.1.0
# langchain-community>=0.0.10

# Langflow (install separately or via pip)
# langflow>=1.0.0

# Development dependencies
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
